# WIP: width aware, pretty print formatter