--- a/src/DocumentFormat.jl
+++ b/src/DocumentFormat.jl
@@ -82,7 +82,6 @@ function apply(text, edit::Edit{UnitRange{Int}})
 end
 
 include("pretty.jl")
-#= include("fdoc.jl") =#
 include("passes.jl")
 include("indents.jl")