There are some kinks but I'd like you folks to have a look. Contrary to the current approach which makes precision edits to the file, this approach outputs a canonical pretty printed version of the file. The canonical version is tuned by the `indent_width` (spaces of an indent) and `max_width`, the maximum width of a line in the file. Current width adjustable types are:

```
MacroCall [x] 
BinaryOpSyntaxCall/BinaryOpCall [x]
Conditional [x]
WhereOpCall [x]
Call [x]
Parameters [x]
TupleH [x]
Vect [x]
InvisBrackets [x]
Braces [x]
Export [x]
Import [x]
Using [x]
```

This also aligns comments based on the indentation context, removes unnecessary newlines, i.e. if you have 5 consecutive newlines somewhere it'll cut it down to 1.

* `pretty` is a recursive function returns an `Edit` (the pretty printed output). 
* `merge_edits` combines two `Edit`s. 
* `newline_ranges` creates an array of line ranges, this is used to gather the start and end lines of an `Edit` which are used when merging to determine whether they should be joined on the same line and separate lines, also it's used collect comments which are in between `Edit`s.
 
### Example outputs

I think this is a good example output to compare:

https://github.com/JuliaDiff/ChainRules.jl/blob/master/src/rules.jl
[output](https://gist.github.com/domluna/bb9ff9ae689b460593831a9eed92e0b3)

This one is a dump of possible combinations
[output](https://gist.github.com/domluna/92bb5db6a5fabac6fee9fc4f7c314976)

I haven't done paid any attention to performance but here is a sample benchmark formatting https://github.com/MikeInnes/MacroTools.jl/blob/master/src/utils.jl.

[output](https://gist.github.com/domluna/ffc84f32faabe7f212957b974ea9a3dd)

```sh
# master
julia> @btime DocumentFormat.format0($s);
  40.233 ms (940472 allocations: 27.90 MiB)

# this branch
julia> @btime DocumentFormat.format($s);
  5.629 ms (40119 allocations: 2.49 MiB)

# functions that don't have anything to do with formatting
julia> @btime CSTParser.parse($s, true);
  1.662 ms (7269 allocations: 381.94 KiB)

julia> @btime collect(CSTParser.Tokenize.tokenize($s));
  580.955 μs (8752 allocations: 538.81 KiB)
```

## TODO

- [ ] More tests
- [ ] Output triple strings properly. I believe this is an issue with CSTParser not labelling the tokens correctly. **This doesn't affect docstrings.**
- [ ] Unicode operator lengths might be > 1, currently they are 1. Example: `lastindex("→") = 1`, while `lastindex("→ ")` = 4.
- [ ] Refactor `pretty` function into 2-3 separate functions.