@davidanthoff 2018-11-22
Thanks for putting so much time into this! Here are some quick reactions:
1) whatever we do, I think we should wait with such a large refactoring until we have shipped the VS Code extension v0.11, and then tackle this for v0.12.
2) I think line breaks could be an excellent optional feature that folks can opt-into, but I'm a bit nervous about making that the default, i.e. I think I would prefer if we could keep the default for VS Code to not change anything about line breaks (neither add nor remove any). An option that allows that to change though would be great!

A more general comment: it would be awesome if at some point we could enable the automatic formatting of a line during editing, i.e. when one starts a new line. I think the LS protocol supports this, and mainly we would have to enable that. I don't know how that would interact with a whole doc approach here, but I think for that feature, we would need a story that allows us to only format a single line.