@davidanthoff 2018-11-22
> Is there an issue with how VSCode handles line breaks?

I'm not clear what you mean by this :) Can you elaborate a bit more?