@domluna 2019-03-24
@davidanthoff @ZacLN I've continued working on this https://github.com/domluna/JLFmt. It works much better now.