@davidanthoff 2019-03-26
Is there a chance that we could eventually make this available here as well somehow? I'm sorry that I can't be of more help, this part of the code is all in @ZacLN's hands :)