@domluna 2019-03-27
@davidanthof Why not just switch to `JLFmt `once the tires are kicked in? Wouldn't just be switching using `DocumentFormat` to using `JLFmt`?