@domluna 2019-03-27
The formatter at JLFmt is completely different than the one here, so merging it back here doesn't make much sense. The actual API is exactly the same though, just call `format` (which is exported). So it should be as simple as

```julia
using JLFmt # instead of DocumentFormat
format(s)
```