@norru 2020-01-15
I'd love to have an option in VSC to switch formatting backend from DocumentFormat to JLFmt/JuliaFormatter!